Refactoring: split into separate scripts & document functions, including expected input and output
